refactor out vscode lm
